updated pipeline, paused for major decision making.
Currently, we are at a decision point where we need to decide if iFrame
extraction should be based on Shell script or Python.

FFMPEG does not have a clean code readily available for Python to extract iFrames,
so the cleaner way is to embed "ListOfVideos -> read into FFMPEG -> create new dir"
into shell script, but it takes time to understand how to perform S3 FileStream
read on shell script.